Format GEMM alternative data to match the WCGOP

Format GEMM alternative data to match the WCGOP

format_gemm_alt(
  dir,
  species_name,
  areas_to_keep,
  gears_to_keep,
  data_file = "SATAN_GEMM_2026-04-12.csv"
)

Arguments

dir

Directory where output will be saved. The directory where the file should be saved. If dir = NULL no output will be saved.

species_name

A string to specify the species name that data should be summarized for. The string should match the name in the alternative gemm data file and needs to be a single species.

areas_to_keep

A vector of area names in the alternative GEMM to filter the data for.

gears_to_keep

A vector of gear names in the alternative GEMM to filter the data for.

data_file

A string of the CSV file name of the alternative GEMM data provided by the FOS team for calculating data weights using alternative areas. The gear type (minor) and the gear group (major) are used to calculate totals by area and weighting proportions.

Value

dataframe

Author

Chantel Wetzel